+++ game/missiongenerator/aircraft/waypoints/splitpoint.py
@@ -1,5 +1,5 @@
 from dcs.point import MovingPoint
-from dcs.task import OptECMUsing
+from dcs.task import OptECMUsing, OptFormation
 
 from .pydcswaypointbuilder import PydcsWaypointBuilder
 
@@ -15,3 +15,5 @@ class SplitPointBuilder(PydcsWaypointBuilder):
             # Let the AI use ECM to defend themselves.
             ecm_option = OptECMUsing(value=OptECMUsing.Values.UseIfOnlyLockByRadar)
             waypoint.tasks.append(ecm_option)
+
+            waypoint.tasks.append(OptFormation.finger_four_close())